creation of new classes and updates
